feat!: Refactoring everything
